Cisco CSCuj22803: fixes for Bsend
changes required to support MPI_Bsend().  Introduces concept of
attaching a buffer to a large segment that the PML can scribble into and
we will send from.  The reason we don't use a pinned buffer and send
directly from that is that usnic_verbs does not (yes) support num_sge>1
for regular sends.  This means the data gets copied twice, but that is
unavoidable.

changed the logic in handle_large_send to be more sensible

Incorporated David's review comments

This commit was SVN r29184.

+++ ompi/mca/btl/usnic/btl_usnic_recv.c
@@ -281,8 +281,9 @@ opal_output(0, "Start PUT to %p\n", chunk_hdr->ch_hdr.put_addr);
 
                 /* Pass this segment up to the PML */
 #if MSGDEBUG2
-                opal_output(0, "  large FRAG complete, pass up %p, %"PRIu64" bytes, tag=%d\n",
-                        desc.des_dst->seg_addr.pval, desc.des_dst->seg_len,
+                opal_output(0, "  large FRAG complete, pass up %p, %u bytes, tag=%d\n",
+                        desc.des_dst->seg_addr.pval,
+                        (unsigned)desc.des_dst->seg_len,
                         (int)chunk_hdr->ch_hdr.tag);
 #endif
                 reg = mca_btl_base_active_message_trigger +
